Create Pyton
This wallet implementation uses the cryptography library to generate a new RSA private key and corresponding public key on initialization. It also provides methods for getting the wallet's address, signing transactions with the private key, and verifying transactions using the public key.

To use this wallet to accept transactions for the token, you would simply create a new Wallet object and provide its public key as the recipient address when creating token transfer transactions. When receiving transactions, you would use the verify_transaction method to ensure that the transaction was signed by a valid sender and has not been tampered with.